Preview e2e speed: split the itx monolith, slowest-first scheduling, 6 workers, bounded spec tails
- apps/os/e2e/vitest/itx.e2e.test.ts (40 tests, 104s wall — the lane's long
  pole) split into 7 thematic files + itx-test-support.ts so file-level
  workers can spread it. Test bodies byte-identical.
- SlowestFirstSequencer: vitest hands files to workers in sort order; a slow
  file starting last tails the whole lane. Observed durations pinned in
  observedFileSeconds.
- maxWorkers 4 -> 6 (peak ~12 concurrent tests; the slot got cheaper per
  test since the 4x2 stopgap — see config comment for history and the
  marathon revalidation rule).
- agent-chat spec: 480s override -> the 240s heavy ceiling, stream waits
  180s -> 90s. With one retry this spec WAS the lane's worst-case tail
  (2x480s); the e2e-policy guard now also scans specs/.
- specs lane wrapped in the same plain watchdog as the vitest lane
  (OS_PREVIEW_VITEST_LANE_TIMEOUT_SECS renamed OS_PREVIEW_LANE_TIMEOUT_SECS)
  — before this a wedged onboarding agent stretched one preview test step
  to 13 minutes (run dpmddk1b75).
