+++ packages/shared/src/test-support/e2e-policy/budgets.ts
@@ -69,12 +69,14 @@ export const E2E_TEST_TIMEOUT_MS = 120_000;
 export const E2E_HEAVY_TEST_TIMEOUT_MS = 240_000;
 
 /**
- * Watchdog on the whole preview vitest lane (`timeout N pnpm e2e` in
- * scripts/preview/preview.ts). Sits just above a healthy lane (itx monolith
- * plus the heavy tests' own caps, running concurrently). Kills, never
- * retries: a lane that blows this is wedged, not slow.
+ * Watchdog on each preview sub-lane (`timeout N pnpm e2e` and
+ * `timeout N pnpm spec` in scripts/preview/preview.ts). Sits well above a
+ * healthy lane (both sub-lanes run ~2-3.5 minutes concurrently). Kills,
+ * never retries: a lane that blows this is wedged, not slow — before the
+ * specs lane was bounded, a wedged onboarding agent stretched one preview
+ * test step to 13 minutes (2026-07-09, run dpmddk1b75).
  */
-export const OS_PREVIEW_VITEST_LANE_TIMEOUT_SECS = 360;
+export const OS_PREVIEW_LANE_TIMEOUT_SECS = 360;
 
 /**
  * Watchdog on one whole preview e2e run (flake-hunt-loop.sh, marathon). A
