+++ packages/shared/src/test-support/e2e-policy/index.ts
@@ -3,7 +3,7 @@ export {
   E2E_CI_RETRY_DELAY_MS,
   E2E_HEAVY_TEST_TIMEOUT_MS,
   E2E_TEST_TIMEOUT_MS,
-  OS_PREVIEW_VITEST_LANE_TIMEOUT_SECS,
+  OS_PREVIEW_LANE_TIMEOUT_SECS,
   PREVIEW_RUN_WATCHDOG_SECS,
   SPEC_ACTION_TIMEOUT_MS,
   SPEC_EXPECT_TIMEOUT_MS,
