+++ specs/agent-chat.spec.ts
@@ -16,8 +16,12 @@ test("onboarding agent replies to a chat message in the feed", async ({
   baseURL,
 }) => {
   // Two full onboarding LLM turns (unprompted greeting can itself be multi-
-  // script) plus UI paint. Under preview load this regularly exceeds 90s/turn.
-  test.setTimeout(480_000);
+  // script) plus UI paint. Under preview load a turn can exceed 90s, so this
+  // gets the heavy-test ceiling (mirrors E2E_HEAVY_TEST_TIMEOUT_MS for the
+  // vitest agent tests) — NOT more: with one retry, this spec's timeout is
+  // the preview lane's worst-case tail (2× this value), and a wedged
+  // onboarding agent should fail loudly, not stretch the lane.
+  test.setTimeout(240_000);
   await using fixture = await helpers.createFixture("agent-chat");
   if (!baseURL) throw new Error("Playwright baseURL fixture is required.");
 
@@ -52,7 +56,7 @@ test("onboarding agent replies to a chat message in the feed", async ({
     const greeting = await agent.stream.waitForEvent({
       afterOffset: 0,
       eventTypes: [WEB_MESSAGE_SENT],
-      timeoutMs: 180_000,
+      timeoutMs: 90_000,
     });
     await page.locator(assistantMessage).first().waitFor({ timeout: 30_000 });
 
@@ -64,7 +68,7 @@ test("onboarding agent replies to a chat message in the feed", async ({
     await agent.stream.waitForEvent({
       afterOffset: greeting.offset,
       eventTypes: [WEB_MESSAGE_SENT],
-      timeoutMs: 180_000,
+      timeoutMs: 90_000,
       predicate: (event) => {
         const text = (event.payload as { message?: unknown } | undefined)?.message;
         return typeof text === "string" && text.includes(marker);
